chore: delete diffUtils.ts, inline createResizingWidget
Moved the only used function into reviewComponents.tsx.
Removed 717 lines of dead diff utility code.
